Fix incorrect behavior of the GPG agent
The GPG_TTY environment variable wasn't being updated every time a new
terminal window was opened so it stayed set to the old value, causing
the agent to fail when trying to use SSH authentication.

The code that resets the variable is now in config.fish and gets updated
every time a terminal is opened.