Remove the explicit keybinding for clang-format
The automatic formatting of the current document would usually be
handled by the LSP client in the editor so there's no need to make
clang-format explicit; not to mention it only makes sense to use it in
C/C++ files, while the `LSP: Format Document` command works for any
project LSP is managing.

Moreover, this command depends on an external script - which on Arch
Linux is provided by the `clang` package - that I symlinked in the
User folder. That wouldn't allow to restore my dotfiles cleanly.

For these reasons, I decided to just remove this keybinding.