Set npm prefix in npmrc file
Settings related to npm are no longer defined as environment variables
but they're stored in a designated file.
Moreover, the node_modules folder where global packages are installed is
no longer in the PATH in order to reduce ambiguity; node packages will
now be invoked with npx, which seems like a better approach to me.