Improve upgrade-system alias
The pattern being matched against has been refined: now it'll only pick
"ID" if it's at the beginning of the line. This change was needed
because in the case of distros such as Ubuntu, the previous pattern
matched multiple strings. This new pattern literally matches only "ID=".

Moreover, a new case has been added for Ubuntu for the upgrade-system
alias, which will expand to the appropriate command.