Fix detection of terminal capabilities
Relying on $DISPLAY was not an appropriate way to detect the terminal
capabilities. The $TERM variable serves the purpose a little better.
Right now I'm only checking for strings that start with xterm. I know
this is not comprehensive but I'll add more strings in case I need them.

Furthermore, in case the terminal doesn't have the capabilities we need,
ed is used as a default instead of nano.