Re-enable dot alias to manage dotfiles
Having a .git directory in my home directory is inconvenient when
running git commands (especially aliases) in subfolders that do not
have another git repository that overshadows the dotfiles repo.
Git will in fact use my home directory as the working directory when
running the alias, which is not desirable.