Remove the `--quit-if-one-screen` option from `less`
This option causes problems with terminal emulators that use an
alternate screen buffer, such as iTerm2. If this option is set, and
`less` is being told to display less than one screen's worth of content,
the user will see no output because `less` will quit immediately in the
alternate buffer.